Fixed rebase introduced errors.
