+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/spring/data/entity/DgvaStudyBrowser.java
@@ -1,7 +1,8 @@
-package uk.ac.ebi.variation.eva.lib.spring.data.entity;
+package uk.ac.ebi.eva.lib.spring.data.entity;
 
-import uk.ac.ebi.variation.eva.lib.datastore.EvaproUtils;
-import uk.ac.ebi.variation.eva.lib.models.VariantStudy;
+
+import uk.ac.ebi.eva.lib.datastore.EvaproUtils;
+import uk.ac.ebi.eva.lib.models.VariantStudy;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
