+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/spring/data/metadata/SpringStudyEvaproDBAdaptor.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.variation.eva.lib.spring.data.metadata;
+package uk.ac.ebi.eva.lib.spring.data.metadata;
 
 import org.opencb.datastore.core.QueryOptions;
 import org.opencb.datastore.core.QueryResult;
@@ -6,40 +6,39 @@ import org.opencb.opencga.storage.core.adaptors.StudyDBAdaptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
-import uk.ac.ebi.variation.eva.lib.datastore.EvaproUtils;
-import uk.ac.ebi.variation.eva.lib.models.VariantStudy;
-import uk.ac.ebi.variation.eva.lib.spring.data.entity.DgvaStudyBrowser;
-import uk.ac.ebi.variation.eva.lib.spring.data.repository.DGvaStudyBrowserRepository;
+import uk.ac.ebi.eva.lib.models.VariantStudy;
+import uk.ac.ebi.eva.lib.spring.data.entity.StudyBrowser;
+import uk.ac.ebi.eva.lib.spring.data.repository.StudyBrowserRepository;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static uk.ac.ebi.variation.eva.lib.spring.data.utils.DgvaDBUtils.getSpeciesAndTypeFilters;
+import static uk.ac.ebi.eva.lib.spring.data.utils.EvaproDbUtils.getSpeciesAndTypeFilters;
+
 
 /**
- * Created by jorizci on 03/10/16.
+ * Created by jorizci on 04/10/16.
  */
 @Component
-public class SpringStudyDgvaDBAdaptor implements StudyDBAdaptor {
+public class SpringStudyEvaproDBAdaptor implements StudyDBAdaptor {
 
     @Autowired
-    private DGvaStudyBrowserRepository dGvaStudyBrowserRepository;
+    private StudyBrowserRepository studyBrowserRepository;
 
     @Override
     public QueryResult getAllStudies(QueryOptions queryOptions) {
         long start = System.currentTimeMillis();
         Specification filterSpecification = getSpeciesAndTypeFilters(queryOptions);
-        List<DgvaStudyBrowser> dgvaStudies = dGvaStudyBrowserRepository.findAll(filterSpecification);
+        List<StudyBrowser> dgvaStudies = studyBrowserRepository.findAll(filterSpecification);
         List<VariantStudy> variantstudies = new ArrayList<>();
-        for (DgvaStudyBrowser dgvaStudy : dgvaStudies) {
-            variantstudies.add(dgvaStudy.generateVariantStudy());
+        for (StudyBrowser dgvaStudy : dgvaStudies) {
+            if (dgvaStudy != null) {
+                variantstudies.add(dgvaStudy.generateVariantStudy());
+            }
         }
         long end = System.currentTimeMillis();
         return new QueryResult(null, ((Long) (end - start)).intValue(), variantstudies.size(), variantstudies.size(), null, null, variantstudies);
     }
-
 
     @Override
     public QueryResult listStudies() {
@@ -52,13 +51,12 @@ public class SpringStudyDgvaDBAdaptor implements StudyDBAdaptor {
     }
 
     @Override
-    public QueryResult getStudyById(String studyId, QueryOptions queryOptions) {
+    public QueryResult getStudyById(String s, QueryOptions queryOptions) {
         long start = System.currentTimeMillis();
-        Specification filterSpecification = getSpeciesAndTypeFilters(queryOptions);
-        DgvaStudyBrowser dgvaStudy = dGvaStudyBrowserRepository.getOne(studyId);
+        StudyBrowser study = studyBrowserRepository.findOne(s);
         List<VariantStudy> variantStudy = new ArrayList<>();
-        if(dgvaStudy!=null){
-            variantStudy.add(dgvaStudy.generateVariantStudy());
+        if (study != null) {
+            variantStudy.add(study.generateVariantStudy());
         }
         long end = System.currentTimeMillis();
         return new QueryResult(null, ((Long) (end - start)).intValue(), variantStudy.size(), variantStudy.size(), null, null, variantStudy);
