+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/spring/data/repository/FileRepository.java
@@ -1,10 +1,9 @@
-package uk.ac.ebi.variation.eva.lib.spring.data.repository;
+package uk.ac.ebi.eva.lib.spring.data.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import uk.ac.ebi.variation.eva.lib.spring.data.ResultClasses.FileFtpReference;
-import uk.ac.ebi.variation.eva.lib.spring.data.entity.File;
+import uk.ac.ebi.eva.lib.spring.data.ResultClasses.FileFtpReference;
+import uk.ac.ebi.eva.lib.spring.data.entity.File;
 
 import java.util.List;
