+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/spring/data/repository/ProjectRepository.java
@@ -1,7 +1,7 @@
-package uk.ac.ebi.variation.eva.lib.spring.data.repository;
+package uk.ac.ebi.eva.lib.spring.data.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uk.ac.ebi.variation.eva.lib.spring.data.entity.Project;
+import uk.ac.ebi.eva.lib.spring.data.entity.Project;
 
 /**
  * Created by jorizci on 03/10/16.
