+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/spring/data/repository/TaxonomyRepository.java
@@ -1,8 +1,8 @@
-package uk.ac.ebi.variation.eva.lib.spring.data.repository;
+package uk.ac.ebi.eva.lib.spring.data.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uk.ac.ebi.variation.eva.lib.models.Assembly;
-import uk.ac.ebi.variation.eva.lib.spring.data.entity.Taxonomy;
+import uk.ac.ebi.eva.lib.models.Assembly;
+import uk.ac.ebi.eva.lib.spring.data.entity.Taxonomy;
 
 import java.util.List;
