+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/spring/data/utils/DgvaDBUtils.java
@@ -1,13 +1,13 @@
-package uk.ac.ebi.variation.eva.lib.spring.data.utils;
+package uk.ac.ebi.eva.lib.spring.data.utils;
 
 import org.opencb.datastore.core.QueryOptions;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.domain.Specifications;
-import uk.ac.ebi.variation.eva.lib.spring.data.repository.DGvaStudyBrowserRepository;
+import uk.ac.ebi.eva.lib.spring.data.repository.DGvaStudyBrowserRepository;
 
 import static org.springframework.data.jpa.domain.Specifications.where;
-import static uk.ac.ebi.variation.eva.lib.spring.data.extension.GenericSpecifications.in;
-import static uk.ac.ebi.variation.eva.lib.spring.data.extension.GenericSpecifications.like;
+import static uk.ac.ebi.eva.lib.spring.data.extension.GenericSpecifications.in;
+import static uk.ac.ebi.eva.lib.spring.data.extension.GenericSpecifications.like;
 
 /**
  * Created by jorizci on 03/10/16.
