+++ eva-server/src/main/java/uk/ac/ebi/eva/server/Application.java
@@ -19,12 +19,15 @@
 
 package uk.ac.ebi.eva.server;
 
+import com.mongodb.MongoClient;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.web.SpringBootServletInitializer;
 import org.springframework.boot.orm.jpa.EntityScan;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.data.mongodb.MongoDbFactory;
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.RequestHandlerSelectors;
@@ -35,6 +38,7 @@ import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 import uk.ac.ebi.eva.lib.datastore.DBAdaptorConnector;
 import uk.ac.ebi.eva.lib.datastore.MultiMongoDbFactory;
+import uk.ac.ebi.eva.lib.spring.data.extension.ExtendedJpaRepositoryFunctionsImpl;
 
 import java.io.IOException;
 import java.util.Properties;
@@ -43,7 +47,7 @@ import java.util.Properties;
 @EnableSwagger2
 @EntityScan(basePackages = {"uk.ac.ebi.eva.lib.spring.data.entity"})
 @EnableJpaRepositories(basePackages = {"uk.ac.ebi.eva.lib.spring.data.repository"}, repositoryBaseClass = ExtendedJpaRepositoryFunctionsImpl.class)
-@ComponentScan(basePackages = {"uk.ac.ebi.eva.lib.spring.data","uk.ac.ebi.variation.eva.server"})
+@ComponentScan(basePackages = {"uk.ac.ebi.eva.lib.spring.data","uk.ac.ebi.eva.server"})
 public class Application extends SpringBootServletInitializer {
 
     @Override
@@ -68,22 +72,6 @@ public class Application extends SpringBootServletInitializer {
         MongoClient mongoClient = DBAdaptorConnector.getMongoClient(properties);
         return new MultiMongoDbFactory(mongoClient, "test");
     }
-
-//    /**
-//     * Create a Datasource bean with the connection to a jndi datasource.
-//     *
-//     * @return
-//     * @throws IOException
-//     * @throws NamingException
-//     */
-//    @Bean(name = "dataSource")
-//    public DataSource evaProDataSource() throws IOException, NamingException {
-//        Properties properties = new Properties();
-//        properties.load(Application.class.getResourceAsStream("/eva.properties"));
-//        String dsName = properties.getProperty("eva.evapro.datasource", "evapro");
-//        JndiTemplate jndiTemplate = new JndiTemplate();
-//        return (DataSource) jndiTemplate.lookup("java:/comp/env/jdbc/" + dsName);
-//    }
 
     @Bean
     public Docket apiConfiguration() {
